+++ fili-core/src/main/java/com/yahoo/bard/webservice/table/PhysicalTable.java
@@ -9,7 +9,6 @@ import com.yahoo.bard.webservice.data.metric.MetricColumn;
 import com.yahoo.bard.webservice.data.time.ZonedTimeGrain;
 import com.yahoo.bard.webservice.metadata.SegmentMetadata;
 import com.yahoo.bard.webservice.util.IntervalUtils;
-import com.yahoo.bard.webservice.util.StreamUtils;
 import com.yahoo.bard.webservice.util.Utils;
 
 import org.joda.time.DateTime;
@@ -18,13 +17,16 @@ import org.joda.time.ReadablePeriod;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.LinkedHashSet;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.atomic.AtomicReference;
+import java.util.stream.Collectors;
 
 import javax.validation.constraints.NotNull;
 
@@ -38,7 +40,7 @@ public class PhysicalTable extends Table {
     private Map<Column, Set<Interval>> workingIntervals;
     private final Object mutex = new Object();
     private final Map<String, String> logicalToPhysicalColumnNames;
-    private final Map<String, String> physicalToLogicalColumnNames;
+    private final Map<String, List<String>> physicalToLogicalColumnNames;
 
     /**
      * Create a physical table.
@@ -58,10 +60,12 @@ public class PhysicalTable extends Table {
         this.workingIntervals = Collections.synchronizedMap(new HashMap<>());
         this.logicalToPhysicalColumnNames = Collections.unmodifiableMap(logicalToPhysicalColumnNames);
         this.physicalToLogicalColumnNames = Collections.unmodifiableMap(
-                this.logicalToPhysicalColumnNames
-                        .entrySet()
-                        .stream()
-                        .collect(StreamUtils.toLinkedMap(Map.Entry::getValue, Map.Entry::getKey))
+                this.logicalToPhysicalColumnNames.entrySet().stream().sequential().collect(
+                        Collectors.groupingBy(
+                                Map.Entry::getValue,
+                                Collectors.mapping(Map.Entry::getKey, Collectors.toList())
+                        )
+                )
         );
     }
 
@@ -137,14 +141,17 @@ public class PhysicalTable extends Table {
             workingIntervals.clear();
             for (Map.Entry<String, Set<Interval>> nameIntervals : dimensionIntervals.entrySet()) {
                 String physicalName = nameIntervals.getKey();
-                String apiName = getLogicalColumnName(physicalName);
-                Dimension dimension = dimensionDictionary.findByApiName(apiName);
-                // Schema evolution may lead to unknown dimensions, skip these
-                if (dimension == null) {
-                    continue;
+                List<String> apiNames = getLogicalColumnName(physicalName);
+
+                for (String apiName : apiNames) {
+                    Dimension dimension = dimensionDictionary.findByApiName(apiName);
+                    // Schema evolution may lead to unknown dimensions, skip these
+                    if (dimension == null) {
+                        continue;
+                    }
+                    DimensionColumn dimensionColumn = DimensionColumn.addNewDimensionColumn(this, dimension);
+                    workingIntervals.put(dimensionColumn, nameIntervals.getValue());
                 }
-                DimensionColumn dimensionColumn = DimensionColumn.addNewDimensionColumn(this, dimension);
-                workingIntervals.put(dimensionColumn, nameIntervals.getValue());
             }
             for (Map.Entry<String, Set<Interval>> nameIntervals : metricIntervals.entrySet()) {
                 MetricColumn metricColumn = MetricColumn.addNewMetricColumn(this, nameIntervals.getKey());
@@ -237,8 +244,8 @@ public class PhysicalTable extends Table {
      * @param physicalName  Physical name to lookup in physical table
      * @return Translated physicalName if applicable
      */
-    private String getLogicalColumnName(String physicalName) {
-        return physicalToLogicalColumnNames.getOrDefault(physicalName, physicalName);
+    private List<String> getLogicalColumnName(String physicalName) {
+        return physicalToLogicalColumnNames.getOrDefault(physicalName, Arrays.asList(physicalName));
     }
 
     /**